modf

+++ App_Restaurant/src/app/components/cart/cart.component.ts
@@ -47,33 +47,46 @@ export class CartComponent {
   }
 
   checkout() {
-    const cartItems=this.cartItems;
-    const username=localStorage.getItem("username");
-    const status ="completed";
-    
-    const order={cartItems,username,status};
+    const list_id_dish: string[] = [];
+    const username = localStorage.getItem("username");
+    const status = "completed";
+  
+    // Construire la liste des IDs des plats en tenant compte des quantités
+    this.cartItems.forEach((item) => {
+      for (let i = 0; i < item.quantity; i++) {
+        list_id_dish.push(item._id); // Utiliser `_id` pour ajouter l'ID des plats
+      }
+    });
+  
+    // Créer l'objet commande
+    const order = { list_id_dish, username, status };
+    console.log("Liste des IDs des plats : ", list_id_dish);
+    console.log("Commande complète : ", order);
+  
+    // Effet de confettis pour la confirmation
     confetti({
-      particleCount: 100, 
-      spread: 70, 
+      particleCount: 100,
+      spread: 70,
       origin: { y: 0.6 }
     });
-
-    this.orderService.sendOrder(this.cartItems).subscribe(
-
-      (response) => {
-        console.log('Command send successfully:', response);
-        this.cartService.removeAllFromCart();
-        this.cartItems = this.cartService.getCartItems();
-        this.cartsubmit=true;
-        setTimeout(() => {
-          this.cartsubmit=false;
-        }, 7000);
-      },
-      (error) => {
-        console.error('Error updating profile:', error);
-      }
-    );
-    
-    
+  
+    // Envoyer la commande via le service
+    if (username){
+      this.orderService.sendOrder(order).subscribe(
+        (response) => {
+          console.log('Commande envoyée avec succès :', response);
+          this.cartService.removeAllFromCart(); // Vider le panier après la commande
+          this.cartItems = this.cartService.getCartItems(); // Rafraîchir le panier
+          this.cartsubmit = true;
+          setTimeout(() => {
+            this.cartsubmit = false;
+          }, 7000);
+        },
+        (error) => {
+          console.error('Erreur lors de l\'envoi de la commande :', error);
+        }
+      );
+    }
   }
+  
 }
